Adds functionality for Cozy Weather 3 to be used with scenes.
There still needs to be some fixes to the Biomes and CozyAtmosphereModule, but once those are resolved, then this will update the FXBlockZones and Reflection Modules.  Other modules can be added with ease without needing to modify them directly from the Cozy package.

Though, ideally, this functionality can be added to Cozy directly to simplify the process.  This is a good start though.